Moved function to zygote utils
