+++ apps/prairielearn/python/prairielearn.py
@@ -1810,24 +1810,3 @@ def index2key(i):
 
 def is_within_limits(n: int) -> bool:
     return -((2**53) - 1) <= n <= 2**53 - 1
-
-
-def all_integers_within_limits(item: Any) -> None:
-    """Raise an exception if the input item contains any oversized integers."""
-    item_stack = [item]
-
-    while item_stack:
-        next_item = item_stack.pop()
-
-        if isinstance(next_item, int):
-            if not is_within_limits(next_item):
-                raise ValueError(
-                    f"Data structure contains oversized integer: {next_item}"
-                )
-
-        elif isinstance(next_item, list):
-            item_stack.extend(next_item)
-
-        elif isinstance(next_item, dict):
-            item_stack.extend(next_item.keys())
-            item_stack.extend(next_item.values())
